+++ ui/app/reducers.js
@@ -4,10 +4,10 @@ const copyToClipboard = require('copy-to-clipboard')
 //
 // Sub-Reducers take in the complete state and return their sub-state
 //
-const reduceIdentities = require('./reducers/identities')
 const reduceMetamask = require('./reducers/metamask')
 const reduceApp = require('./reducers/app')
 const reduceLocale = require('./reducers/locale')
+const reduceSend = require('./ducks/send.duck').default
 
 window.METAMASK_CACHED_LOG_STATE = null
 
@@ -20,12 +20,6 @@ function rootReducer (state, action) {
   if (action.type === 'GLOBAL_FORCE_UPDATE') {
     return action.value
   }
-
-  //
-  // Identities
-  //
-
-  state.identities = reduceIdentities(state, action)
 
   //
   // MetaMask
@@ -44,6 +38,12 @@ function rootReducer (state, action) {
   //
 
   state.localeMessages = reduceLocale(state, action)
+
+  //
+  // Send
+  //
+
+  state.send = reduceSend(state, action)
 
   window.METAMASK_CACHED_LOG_STATE = state
   return state