+++ ui/app/util.js
@@ -291,7 +291,7 @@ function getTokenAddressFromTokenObject (token) {
 
 /**
  * Safely checksumms a potentially-null address
- * 
+ *
  * @param {String} [address] - address to checksum
  * @returns {String} - checksummed address
  */